+++ lib/gpt3-chat-completion/main.go
@@ -21,7 +21,7 @@ type Conversation struct {
 	ID          string `json:"id"`
 	Participant string `json:"participant"`
 	Message     string `json:"message"`
-	Timestamp   int32  `json:"timestamp"`
+	Timestamp   int64  `json:"timestamp"`
 }
 
 func HandleRequest(conversation []Conversation) (events.APIGatewayProxyResponse, error) {
